+++ test/output_c/include/std_shared_ptr_int_array.h
@@ -59,7 +59,7 @@ MR_C_API int *MR_C_std_shared_ptr_int_array_At(const MR_C_std_shared_ptr_int_arr
 /// This being zero usually conincides with `MR_C_std_shared_ptr_int_array_Get()` returning null, but is ultimately orthogonal.
 /// Note that in multithreaded environments, the only safe way to use this number is comparing it with zero. Positive values might change by the time you get to use them.
 /// Parameter `_this` can not be null.
-MR_C_API long MR_C_std_shared_ptr_int_array_UseCount(const MR_C_std_shared_ptr_int_array *_this);
+MR_C_API int MR_C_std_shared_ptr_int_array_UseCount(const MR_C_std_shared_ptr_int_array *_this);
 
 /// Create a new instance, taking ownership of an existing pointer.
 /// Parameter `ptr` can point to an array.
